+++ src/components/SpBlockInfoCard.vue
@@ -40,7 +40,7 @@
 import { mapGetters, mapMutations, mapActions } from 'vuex'
 import { formatter as blockFormatter } from '../helpers/block'
 
-import Box from "../assets/icons/Box"
+import Box from "./icons/Box"
 
 export default {
   components: {
