+++ src/components/icons/Copy.vue
