+++ src/components/icons/IconCircle2.vue
