+++ src/components/icons/IconLoading2.vue
