+++ src/components/icons/IconLock3.vue
