+++ src/components/icons/IconMagic1.vue
