+++ src/components/icons/IconPaperPlane2.vue
