+++ src/components/icons/IconSynchronization1.vue
