+++ src/components/icons/Spinner.vue
